BUG: Remove shadow PR files (#745)
Closes #729 

As detailed
[here](https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-excluding-paths),
the `paths-ignore` keyword in the shadow PR workflows does not work as
we expected:

> When all the path names match patterns in paths-ignore, the workflow
will not run. If any path names do not match patterns in paths-ignore,
even if some path names match the patterns, the workflow will run.

This means that if _any_ files in the PR do not match any of the
`paths-ignore` patterns the shadow workflow will trigger, which is very
much not our desired behaviour. This solution would only work for very
simple PRs editing a single file.

As such, we're stripping out these shadow PRs and simply disabling
auto-merge for now until we think of a better solution / one becomes
available!