BUG: Gather results causes OOM for very large bag_size (#765)
We discovered a bug after merging #754 val_set_is_dist was used as a
flag to gather_results. It was defined as
 val_set_is_dist  = pl_replace_sampler_ssp and max_num_gpus > 1 
However max_num_gpus is set to -1 by default. So val_set_is_dist was
always False.
Now that we deprecated this flag results got gathered. This blows up the
memory when we use very large bag_size (1200).
Therefore we only gather if save_intermediate_outputs is True or if it's
an extra validation epoch.
A test is added to make sure gather_results is called when expected only