FIX: Set evaluation mode to compute number of dimensions (#809)
When guessing the number of output dimensions in the image encoder, the
model is in training mode and the batch norm stats get updated. This
shouldn't really happen and makes the checks in
https://github.com/microsoft/hi-ml/pull/807/files#r1112200449 fail. This
PR adds a line to ensure the encode is in `eval` mode before calculating
the number of output dims.