FIX: Use different image for multimodal notebook (#808)
Fixes #805.

Results with the new image:

![image](https://user-images.githubusercontent.com/12688084/220152485-418e0fb0-0e74-4e45-b399-8295b16ec8d7.png)

This image is very large, though :`(3950, 3300)`. Feel free to propose a
smaller one from the internet. Perhaps from Open-i.

<!--
## Guidelines

Please follow the guidelines for pull requests (PRs) in
[CONTRIBUTING](/CONTRIBUTING.md). Checklist:

- Ensure that your PR is small, and implements one change
- Give your PR title one of the prefixes listed in
[CONTRIBUTING](../docs/source/coding_guidelines.md#pull-request-titles)
to indicate what type of change that is
- Link the correct GitHub issue for tracking
- Add unit tests for all functions that you introduced or modified
- Run automatic code formatting / linting on all files ("Format
Document" Shift-Alt-F in VSCode)
- Ensure that documentation renders correctly in Sphinx by running `make
html` in the `docs` folder

## Change the default merge message

When completing your PR, you will be asked for a title and an optional
extended description. By default, the extended description will be a
concatenation of the individual
commit messages. Please DELETE/REPLACE that with a human readable
extended description for non-trivial PRs.
-->

---------

Co-authored-by: Stephanie Hyland <corcra@users.noreply.github.com>