ENH: Add BioViL-T image and text encoders (#830)
- Splits the content in `image/model/model.py` into separate files with
respect to their functionality, such as encoders and custom types.
- Creates a new encoder class that can handle multiple image inputs such
as current scan and previous scan. This way models can embed
multi-series data into a single vector representation.

<!--
## Guidelines

Please follow the guidelines for pull requests (PRs) in
[CONTRIBUTING](/CONTRIBUTING.md). Checklist:

- Ensure that your PR is small, and implements one change
- Give your PR title one of the prefixes listed in
[CONTRIBUTING](../docs/source/coding_guidelines.md#pull-request-titles)
to indicate what type of change that is
- Link the correct GitHub issue for tracking
- Add unit tests for all functions that you introduced or modified
- Run automatic code formatting / linting on all files ("Format
Document" Shift-Alt-F in VSCode)
- Ensure that documentation renders correctly in Sphinx by running `make
html` in the `docs` folder

## Change the default merge message

When completing your PR, you will be asked for a title and an optional
extended description. By default, the extended description will be a
concatenation of the individual
commit messages. Please DELETE/REPLACE that with a human readable
extended description for non-trivial PRs.
-->

---------

Co-authored-by: Stephanie Hyland <sthyland@microsoft.com>
Co-authored-by: Stephanie Hyland <corcra@users.noreply.github.com>
Co-authored-by: Fernando Pérez-García <fepegar@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>