ENH: Add support for other hyperdrive sweep types (#959)
Currently submitting a hyperdrive run assumes that all sweep
distributions are `Choice`. This PR adds support for other sweep types
described
[here](https://learn.microsoft.com/en-us/python/api/azure-ai-ml/azure.ai.ml.sweep?view=azure-python)
for SDK v2 jobs.

If hydra is used for command line arguments then `/` or `.` will be in
the argument names which is not supported as an input type or argument
name for `command_job`. As such, these get replaced with `_`.